requirements.txt

autopep8
azure-storage-blob
click
colorama
Flask
Flask-Cors
flask-cors
Flask-JWT-Extended
Flask-Login
Flask-WTF
importlib-metadata
itsdangerous
Jinja2
MarkupSafe
mlflow
numpy
Pillow
psycopg2
psycopg2-binary
pycodestyle
pyodbc
python-decouple
python-dotenv
six
tensorflow
toml
Werkzeug
WTForms
zipp
